hallo

+++ database/migrations/2021_12_27_172315_create_pembayarans_table.php
@@ -16,7 +16,7 @@ class CreatePembayaransTable extends Migration
         Schema::create('pembayarans', function (Blueprint $table) {
             $table->id();
             $table->date('tgl_bayar');
-            $table->integer('total');
+            $table->bigInteger('total');
             $table->string('metode');
             $table->bigInteger('id_penjualan')->unsigned();
             //relasi
